requirements.txt
South
python-tastypie
